refactor(db): restructure and separate recording and transformation models into dedicated files for improved organization and clarity
