--- apps/app/src/lib/services/db/models.ts
@@ -1,225 +1,0 @@
-import type { TaggedError } from '@epicenterhq/result';
-import type {
-	ANTHROPIC_INFERENCE_MODELS,
-	GOOGLE_INFERENCE_MODELS,
-	GROQ_INFERENCE_MODELS,
-	INFERENCE_PROVIDERS,
-	OPENAI_INFERENCE_MODELS,
-} from '$lib/constants/inference';
-import { nanoid } from 'nanoid/non-secure';
-
-export type DbServiceError = TaggedError<'DbServiceError'>;
-
-export function generateDefaultTransformation(): Transformation {
-	const now = new Date().toISOString();
-	return {
-		id: nanoid(),
-		title: '',
-		description: '',
-		steps: [],
-		createdAt: now,
-		updatedAt: now,
-	};
-}
-
-export function generateDefaultTransformationStep(): TransformationStep {
-	return {
-		id: nanoid(),
-		type: 'prompt_transform',
-		'prompt_transform.inference.provider': 'Google',
-		'prompt_transform.inference.provider.OpenAI.model': 'gpt-4o',
-		'prompt_transform.inference.provider.Groq.model': 'llama-3.3-70b-versatile',
-		'prompt_transform.inference.provider.Anthropic.model': 'claude-sonnet-4-0',
-		'prompt_transform.inference.provider.Google.model': 'gemini-2.5-flash',
-
-		'prompt_transform.systemPromptTemplate': '',
-		'prompt_transform.userPromptTemplate': '',
-
-		'find_replace.findText': '',
-		'find_replace.replaceText': '',
-		'find_replace.useRegex': false,
-	};
-}
-
-export const TRANSFORMATION_STEP_TYPES = [
-	'prompt_transform',
-	'find_replace',
-] as const;
-
-export const TRANSFORMATION_STEP_TYPES_TO_LABELS = {
-	prompt_transform: 'Prompt Transform',
-	find_replace: 'Find Replace',
-} as const satisfies Record<(typeof TRANSFORMATION_STEP_TYPES)[number], string>;
-
-export type Recording = {
-	id: string;
-	title: string;
-	subtitle: string;
-	timestamp: string;
-	createdAt: string;
-	updatedAt: string;
-	transcribedText: string;
-	blob: Blob | undefined;
-	/**
-	 * A recording
-	 * 1. Begins in an 'UNPROCESSED' state
-	 * 2. Moves to 'TRANSCRIBING' while the audio is being transcribed
-	 * 3. Finally is marked as 'DONE' when the transcription is complete.
-	 * 4. If the transcription fails, it is marked as 'FAILED'
-	 */
-	transcriptionStatus: 'UNPROCESSED' | 'TRANSCRIBING' | 'DONE' | 'FAILED';
-};
-
-export type Transformation = {
-	id: string;
-	title: string;
-	description: string;
-	createdAt: string;
-	updatedAt: string;
-	/**
-	 * It can be one of several types of text transformations:
-	 * - find_replace: Replace text patterns with new text
-	 * - prompt_transform: Use AI to transform text based on prompts
-	 */
-	steps: {
-		id: string;
-		// For now, steps don't need titles or descriptions. They can be computed from the type as "Find and Replace" or "Prompt Transform"
-		type: (typeof TRANSFORMATION_STEP_TYPES)[number];
-
-		'prompt_transform.inference.provider': (typeof INFERENCE_PROVIDERS)[number];
-		'prompt_transform.inference.provider.OpenAI.model': (typeof OPENAI_INFERENCE_MODELS)[number];
-		'prompt_transform.inference.provider.Groq.model': (typeof GROQ_INFERENCE_MODELS)[number];
-		'prompt_transform.inference.provider.Anthropic.model': (typeof ANTHROPIC_INFERENCE_MODELS)[number];
-		'prompt_transform.inference.provider.Google.model': (typeof GOOGLE_INFERENCE_MODELS)[number];
-
-		'prompt_transform.systemPromptTemplate': string;
-		'prompt_transform.userPromptTemplate': string;
-
-		'find_replace.findText': string;
-		'find_replace.replaceText': string;
-		'find_replace.useRegex': boolean;
-	}[];
-};
-
-export type TransformationStep = Transformation['steps'][number];
-export type InsertTransformationStep = Omit<
-	TransformationStep,
-	'createdAt' | 'updatedAt'
->;
-
-/**
- * Base properties shared by all transformation step run variants.
- */
-type BaseTransformationStepRun = {
-	id: string;
-	stepId: string;
-	startedAt: string;
-	completedAt: string | null;
-	input: string;
-};
-
-export type TransformationStepRunRunning = BaseTransformationStepRun & {
-	status: 'running';
-};
-
-export type TransformationStepRunCompleted = BaseTransformationStepRun & {
-	status: 'completed';
-	output: string;
-};
-
-export type TransformationStepRunFailed = BaseTransformationStepRun & {
-	status: 'failed';
-	error: string;
-};
-
-export type TransformationStepRun =
-	| TransformationStepRunRunning
-	| TransformationStepRunCompleted
-	| TransformationStepRunFailed;
-
-/**
- * Base properties shared by all transformation run variants.
- *
- * Status transitions:
- * 1. 'running' - Initial state when created and transformation is immediately invoked
- * 2. 'completed' - When all steps have completed successfully
- * 3. 'failed' - If any step fails or an error occurs
- */
-type BaseTransformationRun = {
-	id: string;
-	transformationId: string;
-	/**
-	 * Recording id if the transformation is invoked on a recording.
-	 * Null if the transformation is invoked on arbitrary text input.
-	 */
-	recordingId: string | null;
-	startedAt: string;
-	completedAt: string | null;
-	/**
-	 * Because the recording's transcribedText can change after invoking,
-	 * we store a snapshot of the transcribedText at the time of invoking.
-	 */
-	input: string;
-	stepRuns: TransformationStepRun[];
-};
-
-export type TransformationRunRunning = BaseTransformationRun & {
-	status: 'running';
-};
-
-export type TransformationRunCompleted = BaseTransformationRun & {
-	status: 'completed';
-	output: string;
-};
-
-export type TransformationRunFailed = BaseTransformationRun & {
-	status: 'failed';
-	error: string;
-};
-
-/**
- * Represents an execution of a transformation, which can be run on either
- * a recording's transcribed text or arbitrary input text.
- */
-export type TransformationRun =
-	| TransformationRunRunning
-	| TransformationRunCompleted
-	| TransformationRunFailed;
-
-// Type guards for TransformationRun
-export function isTransformationRunCompleted(
-	run: TransformationRun,
-): run is TransformationRunCompleted {
-	return run.status === 'completed';
-}
-
-export function isTransformationRunFailed(
-	run: TransformationRun,
-): run is TransformationRunFailed {
-	return run.status === 'failed';
-}
-
-export function isTransformationRunRunning(
-	run: TransformationRun,
-): run is TransformationRunRunning {
-	return run.status === 'running';
-}
-
-// Type guards for TransformationStepRun
-export function isTransformationStepRunCompleted(
-	stepRun: TransformationStepRun,
-): stepRun is TransformationStepRunCompleted {
-	return stepRun.status === 'completed';
-}
-
-export function isTransformationStepRunFailed(
-	stepRun: TransformationStepRun,
-): stepRun is TransformationStepRunFailed {
-	return stepRun.status === 'failed';
-}
-
-export function isTransformationStepRunRunning(
-	stepRun: TransformationStepRun,
-): stepRun is TransformationStepRunRunning {
-	return stepRun.status === 'running';
-}
