+++ apps/app/src/lib/services/db/models/recordings.ts
@@ -1,17 +1,30 @@
-import type { Recording, Transformation, TransformationRun } from './models';
+export type Recording = {
+	id: string;
+	title: string;
+	subtitle: string;
+	timestamp: string;
+	createdAt: string;
+	updatedAt: string;
+	transcribedText: string;
+	blob: Blob | undefined;
+	/**
+	 * A recording
+	 * 1. Begins in an 'UNPROCESSED' state
+	 * 2. Moves to 'TRANSCRIBING' while the audio is being transcribed
+	 * 3. Finally is marked as 'DONE' when the transcription is complete.
+	 * 4. If the transcription fails, it is marked as 'FAILED'
+	 */
+	transcriptionStatus: 'UNPROCESSED' | 'TRANSCRIBING' | 'DONE' | 'FAILED';
+};
 
 export type RecordingsDbSchemaV5 = {
 	recordings: Omit<Recording, 'blob'> & {
 		serializedAudio: { arrayBuffer: ArrayBuffer; blobType: string } | undefined;
 	};
-	transformations: Transformation;
-	transformationRuns: TransformationRun;
 };
 
 export type RecordingsDbSchemaV4 = {
 	recordings: Recording;
-	transformations: Transformation;
-	transformationRuns: TransformationRun;
 };
 
 export type RecordingsDbSchemaV3 = {
